requirements.txt
Django==2.2
psycopg2==2.8.5
djangorestframework==3.9.4